backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arc"
version = "1.0.0"
description = "Arc - Autonomous Red Team & Penetration Testing Framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "Arc Team"},
]
keywords = ["pentesting", "red-team", "security", "ai", "agents"]

dependencies = [
    # Web Framework
    "fastapi==0.109.2",
    "uvicorn[standard]==0.27.1",
    "python-multipart==0.0.9",
    "websockets==12.0",

    # Configuration & Validation
    "pydantic==2.6.1",
    "pydantic-settings==2.2.1",
    "email-validator>=2.0.0",

    # Database
    "neo4j==5.17.0",
    "redis==5.0.1",

    # PostgreSQL (async)
    "asyncpg==0.29.0",
    "sqlalchemy[asyncio]==2.0.27",
    "alembic==1.13.1",

    # Qdrant Vector Store
    "qdrant-client==1.7.3",

    # Authentication
    "python-jose[cryptography]==3.3.0",
    "bcrypt>=4.0.0",

    # HTTP Client
    "httpx==0.26.0",
    "aiohttp==3.9.3",

    # Logging
    "structlog==24.1.0",
    "python-json-logger==2.0.7",

    # LLM Providers
    "openai==1.12.0",
    "anthropic==0.18.1",

    # Agent Orchestration (LangGraph)
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",

    # Utilities
    "python-dotenv==1.0.1",
    "tenacity==8.2.3",
    "orjson==3.9.14",

    # Async utilities
    "asyncio-throttle==1.0.2",
    "aiocache==0.12.2",

    # Scheduling
    "apscheduler==3.10.4",

    # Date/Time
    "python-dateutil==2.8.2",

    # Network utilities
    "dnspython==2.6.1",
    "validators==0.22.0",

    # Elasticsearch
    "elasticsearch[async]==8.12.1",

    # GraphQL
    "strawberry-graphql[fastapi]==0.219.2",

    # PDF Export
    "weasyprint==62.3",
    "markdown==3.5.2",

    # CLI
    "click==8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx",                   # for TestClient
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
arc = "cli:main"

# =============================================================================
# Tool Configuration
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring external services",
]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "RUF012", # mutable class variables (used for config)
]

[tool.ruff.lint.isort]
known-first-party = ["core", "agents", "memory", "intelligence", "graph", "recon", "reporting", "api"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
